Create different ViewLoader classes (inc. one for subview).
